finish renaming
