+++ ext/opentelemetry-ext-docker-tests/tests/sqlalchemy_tests/test_postgres.py
@@ -20,7 +20,7 @@ import pytest
 from sqlalchemy.exc import ProgrammingError
 
 from opentelemetry import trace
-from opentelemetry.instrumentation.sqlalchemy.engine import (
+from opentelemetry.ext.sqlalchemy.engine import (
     _DB,
     _HOST,
     _PORT,
