+++ ext/opentelemetry-ext-sqlalchemy/setup.cfg
@@ -55,4 +55,4 @@ where = src
 
 [options.entry_points]
 opentelemetry_instrumentor =
-    sqlalchemy = opentelemetry.instrumentation.sqlalchemy:SQLAlchemyInstrumentor
+    sqlalchemy = opentelemetry.ext.sqlalchemy:SQLAlchemyInstrumentor
