+++ ext/opentelemetry-ext-sqlalchemy/setup.py
@@ -17,12 +17,7 @@ import setuptools
 
 BASE_DIR = os.path.dirname(__file__)
 VERSION_FILENAME = os.path.join(
-    BASE_DIR,
-    "src",
-    "opentelemetry",
-    "instrumentation",
-    "sqlalchemy",
-    "version.py",
+    BASE_DIR, "src", "opentelemetry", "ext", "sqlalchemy", "version.py",
 )
 PACKAGE_INFO = {}
 with open(VERSION_FILENAME) as f:
